Add use client
