+++ demo/src/components/radio.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React, { useCallback, forwardRef } from 'react'
 import { useField } from 'formik'
 import { clsx } from 'cva'
