+++ demo/src/components/sortable.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React, { useState } from 'react'
 import {
   DndContext,
