+++ demo/src/components/switch.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React, { forwardRef } from 'react'
 import { useField } from 'formik'
 import { Label } from './label'
